+++ src/geo/projection/globe_covering_tiles.ts
@@ -1,9 +1,10 @@
-import {vec2} from 'gl-matrix';
-import {IntersectionResult} from '../../util/primitives';
+import {vec2, vec3, vec4} from 'gl-matrix';
+import {Aabb, Frustum, IntersectionResult} from '../../util/primitives';
 import {CoveringTilesOptions} from '../transform_interface';
 import {OverscaledTileID} from '../../source/tile_id';
 import {MercatorCoordinate} from '../mercator_coordinate';
-import type {GlobeTransform} from './globe_transform';
+import {EXTENT} from '../../data/extent';
+import {projectTileCoordinatesToSphere} from './globe_utils';
 
 type CoveringTilesResult = {
     tileID: OverscaledTileID;
@@ -83,6 +84,7 @@ function shouldSplitTile(centerCoord: MercatorCoordinate, cameraCoord: MercatorC
     return Math.min(centerDist, cameraDist) * 2 <= radiusOfMaxLvlLodInTiles; // Multiply distance by 2, because the subdivided tiles would be half the size
 }
 
+// Returns the wrap value for a given tile, computed so that tiles will remain loaded when crossing the antimeridian.
 function getWrap(centerCoord: MercatorCoordinate, tileX: number, tileSize: number): number {
     const distanceCurrent = distanceToTileSimple(centerCoord.x, tileX, tileSize);
     const distanceLeft = distanceToTileSimple(centerCoord.x, tileX - 1.0, tileSize);
@@ -98,14 +100,123 @@ function getWrap(centerCoord: MercatorCoordinate, tileX: number, tileSize: numbe
 }
 
 /**
+ * Returns the AABB of the specified tile. The AABB is in the coordinate space where the globe is a unit sphere.
+ * @param tileID - Tile x, y and z for zoom.
+ */
+export function getTileAABB(tileID: {x: number; y: number; z: number}): Aabb {
+    // We can get away with only checking the 4 tile corners for AABB construction, because for any tile of zoom level 2 or higher
+    // it holds that the extremes (minimal or maximal value) of X, Y or Z coordinates must lie in one of the tile corners.
+    //
+    // To see why this holds, consider the formula for computing X,Y and Z from angular coordinates.
+    // It goes something like this:
+    //
+    // X = sin(lng) * cos(lat)
+    // Y = sin(lat)
+    // Z = cos(lng) * cos(lat)
+    //
+    // Note that a tile always covers a continuous range of lng and lat values,
+    // and that tiles that border the mercator north/south edge are assumed to extend all the way to the poles.
+    //
+    // We will consider each coordinate separately and show that an extreme must always lie in a tile corner for every axis, and must not lie inside the tile.
+    //
+    // For Y, it is clear that the only way for an extreme to not lie on an edge of the lat range is for the range to contain lat=90° or lat=-90° without either being the tile edge.
+    // This cannot happen for any tile, these latitudes will always:
+    // - either lie outside the tile entirely, thus Y will be monotonically increasing or decreasing across the entire tile, thus the extreme must lie at a corner/edge
+    // - or be the tile edge itself, thus the extreme will lie at the tile edge
+    //
+    // For X, considering only longitude, the tile would also have to contain lng=90° or lng=-90° (with neither being the tile edge) for the extreme to not lie on a tile edge.
+    // This can only happen at zoom levels 0 and 1, which are handled separately.
+    // But X is also scaled by cos(lat)! However, this can only cause an extreme to lie inside the tile if the tile crosses lat=0°, which cannot happen for zoom levels other than 0.
+    //
+    // For Z, similarly to X, the extremes must lie at lng=0° or lng=180°, but for zoom levels other than 0 these cannot lie inside the tile. Scaling by cos(lat) has the same effect as with the X axis.
+    //
+    // So checking the 4 tile corners only fails for tiles with zoom level <2, and these are handled separately with hardcoded AABBs:
+    // - zoom level 0 tile is the entire sphere
+    // - zoom level 1 tiles are "quarters of a sphere"
+
+    if (tileID.z <= 0) {
+        // Tile covers the entire sphere.
+        return new Aabb(
+            [-1, -1, -1],
+            [1, 1, 1]
+        );
+    } else if (tileID.z === 1) {
+        // Tile covers a quarter of the sphere.
+        // X is 1 at lng=E90°
+        // Y is 1 at **north** pole
+        // Z is 1 at null island
+        return new Aabb(
+            [tileID.x === 0 ? -1 : 0, tileID.y === 0 ? 0 : -1, -1],
+            [tileID.x === 0 ? 0 : 1, tileID.y === 0 ? 1 : 0, 1]
+        );
+    } else {
+        // Compute AABB using the 4 corners.
+
+        const corners = [
+            projectTileCoordinatesToSphere(0, 0, tileID),
+            projectTileCoordinatesToSphere(EXTENT, 0, tileID),
+            projectTileCoordinatesToSphere(EXTENT, EXTENT, tileID),
+            projectTileCoordinatesToSphere(0, EXTENT, tileID),
+        ];
+
+        const min: vec3 = [1, 1, 1];
+        const max: vec3 = [-1, -1, -1];
+
+        for (const c of corners) {
+            for (let i = 0; i < 3; i++) {
+                min[i] = Math.min(min[i], c[i]);
+                max[i] = Math.max(max[i], c[i]);
+            }
+        }
+
+        // Special handling of poles - we need to extend the tile AABB
+        // to include the pole for tiles that border mercator north/south edge.
+        if (tileID.y === 0 || (tileID.y === (1 << tileID.z) - 1)) {
+            const pole = [0, tileID.y === 0 ? 1 : -1, 0];
+            for (let i = 0; i < 3; i++) {
+                min[i] = Math.min(min[i], pole[i]);
+                max[i] = Math.max(max[i], pole[i]);
+            }
+        }
+
+        return new Aabb(
+            min,
+            max
+        );
+    }
+}
+
+/**
+ * A simple/heuristic function that returns whether the tile is visible under the current transform.
+ * @returns 0 is not visible, 1 if partially visible, 2 if fully visible.
+ */
+function isTileVisible(frustum: Frustum, plane: vec4, x: number, y: number, z: number): IntersectionResult {
+    const tileID = {x, y, z};
+    const aabb = getTileAABB(tileID);
+
+    const frustumTest = aabb.intersectsFrustum(frustum);
+    const planeTest = aabb.intersectsPlane(plane);
+
+    if (frustumTest === IntersectionResult.None || planeTest === IntersectionResult.None) {
+        return IntersectionResult.None;
+    }
+
+    if (frustumTest === IntersectionResult.Full && planeTest === IntersectionResult.Full) {
+        return IntersectionResult.Full;
+    }
+
+    return IntersectionResult.Partial;
+}
+
+/**
  * Returns a list of tiles that optimally covers the screen. Adapted for globe projection.
  * Correctly handles LOD when moving over the antimeridian.
  * @param transform - The globe transform instance.
  * @param options - Additional coveringTiles options.
  * @returns A list of tile coordinates, ordered by ascending distance from camera.
  */
-export function globeCoveringTiles(transform: GlobeTransform, options: CoveringTilesOptions): OverscaledTileID[] {
-    let z = transform.coveringZoomLevel(options);
+export function globeCoveringTiles(frustum: Frustum, plane: vec4, cameraCoord: MercatorCoordinate, centerCoord: MercatorCoordinate, coveringZoom: number, options: CoveringTilesOptions): OverscaledTileID[] {
+    let z = coveringZoom;
     const actualZ = z;
 
     if (options.minzoom !== undefined && z < options.minzoom) {
@@ -115,8 +226,6 @@ export function globeCoveringTiles(transform: GlobeTransform, options: CoveringT
         z = options.maxzoom;
     }
 
-    const cameraCoord = transform.screenPointToMercatorCoordinate(transform.getCameraPoint());
-    const centerCoord = MercatorCoordinate.fromLngLat(transform.center);
     const numTiles = Math.pow(2, z);
     const cameraPoint = [numTiles * cameraCoord.x, numTiles * cameraCoord.y, 0];
     const centerPoint = [numTiles * centerCoord.x, numTiles * centerCoord.y, 0];
@@ -143,7 +252,7 @@ export function globeCoveringTiles(transform: GlobeTransform, options: CoveringT
 
         // Visibility of a tile is not required if any of its ancestor if fully visible
         if (!fullyVisible) {
-            const intersectResult = transform.isTileVisible(it.x, it.y, it.zoom);
+            const intersectResult = isTileVisible(frustum, plane, it.x, it.y, it.zoom);
 
             if (intersectResult === IntersectionResult.None)
                 continue;
