+++ src/util/test/util.ts
@@ -8,6 +8,7 @@ import {MercatorTransform} from '../../geo/projection/mercator_transform';
 import {RequestManager} from '../request_manager';
 import {IReadonlyTransform, ITransform} from '../../geo/transform_interface';
 import {Style} from '../../style/style';
+import type {GlobeProjection} from '../../geo/projection/globe';
 
 export class StubMap extends Evented {
     style: Style;
@@ -222,4 +223,20 @@ export function expectToBeCloseToArray(actual: Array<number>, expected: Array<nu
     for (let i = 0; i < expected.length; i++) {
         expect(actual[i]).toBeCloseTo(expected[i], precision);
     }
+}
+
+export function getGlobeProjectionMock(): GlobeProjection {
+    return {
+        get useGlobeControls(): boolean {
+            return true;
+        },
+        get useGlobeRendering(): boolean {
+            return true;
+        },
+        set useGlobeRendering(_value: boolean) {
+            // do not set
+        },
+        latitudeErrorCorrectionRadians: 0,
+        errorQueryLatitudeDegrees: 0,
+    } as GlobeProjection;
 }
